+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 * Make library works with GHC 8.6 and 8.8 again.
 * Add `pinPrimVector` and `pinPrimArray` to `Z.Foreign`.
+* Export `fail'` from `Z.Data.Parser`
 
 ## 0.1.3.1  -- 2020-09-24